feat(AppInstalling): Add script for automated package and application installation
The `Install-PackagesAndApps.ps1` file was added to automate the installation of packages and applications using Scoop and Winget. The script checks whether tools and applications are already installed and applies necessary actions if they are not. It also sets execution policies for Scoop. The installed programs include Git, Oh My Posh, Neovim, Flow Launcher, Python, and Node.js via fnm.

- Automated installation of applications with Winget.
- Pre-installation checks to avoid redundant installations.
- Refresh of environment variables and PowerShell profile reload after installations.

This implementation streamlines the setup process for essential development tools.